requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==4.0.3
djangorestframework==3.13.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
